split tests into new test file
